+++ phonopy/interface/calculator.py
@@ -36,8 +36,10 @@
 from __future__ import annotations
 
 import pathlib
+import warnings
 from argparse import ArgumentParser
 from collections.abc import Sequence
+from math import pi, sqrt
 from typing import Optional, Union
 
 import numpy as np
@@ -46,26 +48,10 @@ import yaml
 from phonopy.file_IO import get_supported_file_extensions_for_compression
 from phonopy.interface.phonopy_yaml import PhonopyYaml
 from phonopy.interface.vasp import sort_positions_by_symbols
+from phonopy.physical_units import get_physical_units
 from phonopy.structure.atoms import PhonopyAtoms
 from phonopy.structure.cells import determinant
 from phonopy.structure.dataset import get_displacements_and_forces
-from phonopy.units import (
-    AbinitToTHz,
-    Bohr,
-    CastepToTHz,
-    CP2KToTHz,
-    CrystalToTHz,
-    DftbpToTHz,
-    ElkToTHz,
-    FleurToTHz,
-    Hartree,
-    PwscfToTHz,
-    Rydberg,
-    SiestaToTHz,
-    TurbomoleToTHz,
-    VaspToTHz,
-    Wien2kToTHz,
-)
 
 calculator_info = {
     "abacus": {"option": {"name": "--abacus", "help": "Invoke ABACUS mode"}},
@@ -118,8 +104,8 @@ def get_interface_mode(args_dict):
 def convert_crystal_structure(filename_in, interface_in, filename_out, interface_out):
     """Convert crystal structures between different calculator interfaces."""
     cell, _ = read_crystal_structure(filename=filename_in, interface_mode=interface_in)
-    units_in = get_default_physical_units(interface_in)
-    units_out = get_default_physical_units(interface_out)
+    units_in = get_calculator_physical_units(interface_in)
+    units_out = get_calculator_physical_units(interface_out)
     factor = units_in["distance_to_A"] / units_out["distance_to_A"]
     cell.cell = cell.cell * factor
     write_crystal_structure(filename_out, cell, interface_mode=interface_out)
@@ -604,7 +590,20 @@ def get_default_displacement_distance(interface_mode):
     return displacement_distance
 
 
-def get_default_physical_units(interface_mode=None):
+def get_default_physical_units(interface_mode=None) -> dict:
+    """Replace get_calculator_get_physical_units()."""
+    warnings.warn(
+        (
+            "get_default_physical_units is deprecated. "
+            "Use get_calculator_physical_units instead."
+        ),
+        DeprecationWarning,
+        stacklevel=2,
+    )
+    return get_calculator_physical_units(interface_mode=interface_mode)
+
+
+def get_calculator_physical_units(interface_mode=None) -> dict:
     """Return physical units of eachi calculator.
 
     Physical units: energy,  distance,  atomic mass, force,        force constants
@@ -638,91 +637,154 @@ def get_default_physical_units(interface_mode=None):
         "length_unit": None,
         "force_unit": None,
     }
+    physical_units = get_physical_units()
 
     if interface_mode is None or interface_mode in ("vasp", "aims", "lammps", "pwmat"):
+        VaspToTHz = physical_units.DefaultToTHz  # [THz] 15.633302
         units["factor"] = VaspToTHz
-        units["nac_factor"] = Hartree * Bohr
+        units["nac_factor"] = physical_units.Hartree * physical_units.Bohr
         units["distance_to_A"] = 1.0
         units["force_constants_unit"] = "eV/angstrom^2"
         units["length_unit"] = "angstrom"
         units["force_unit"] = "eV/angstrom"
     elif interface_mode == "abinit":
+        AbinitToTHz = (
+            sqrt(physical_units.EV / (physical_units.AMU * physical_units.Bohr))
+            / physical_units.Angstrom
+            / (2 * pi)
+            / 1e12
+        )  # [THz] 21.49068
         units["factor"] = AbinitToTHz
-        units["nac_factor"] = Hartree / Bohr
-        units["distance_to_A"] = Bohr
+        units["nac_factor"] = physical_units.Hartree / physical_units.Bohr
+        units["distance_to_A"] = physical_units.Bohr
         units["force_constants_unit"] = "eV/angstrom.au"
         units["length_unit"] = "au"
         units["force_unit"] = "eV/angstrom"
     elif interface_mode == "qe":
+        PwscfToTHz = (
+            sqrt(physical_units.Rydberg * physical_units.EV / physical_units.AMU)
+            / (physical_units.Bohr * 1e-10)
+            / (2 * pi)
+            / 1e12
+        )  # [THz] 108.97077
         units["factor"] = PwscfToTHz
         units["nac_factor"] = 2.0
-        units["distance_to_A"] = Bohr
-        units["force_to_eVperA"] = Rydberg / Bohr
+        units["distance_to_A"] = physical_units.Bohr
+        units["force_to_eVperA"] = physical_units.Rydberg / physical_units.Bohr
         units["force_constants_unit"] = "Ry/au^2"
         units["length_unit"] = "au"
         units["force_unit"] = "Ry/au"
     elif interface_mode == "wien2k":
+        Wien2kToTHz = (
+            sqrt(physical_units.Rydberg / 1000 * physical_units.EV / physical_units.AMU)
+            / (physical_units.Bohr * 1e-10)
+            / (2 * pi)
+            / 1e12
+        )  # [THz] 3.44595837
         units["factor"] = Wien2kToTHz
         units["nac_factor"] = 2000.0
-        units["distance_to_A"] = Bohr
-        units["force_to_eVperA"] = 0.001 * Rydberg / Bohr
+        units["distance_to_A"] = physical_units.Bohr
+        units["force_to_eVperA"] = 0.001 * physical_units.Rydberg / physical_units.Bohr
         units["force_constants_unit"] = "mRy/au^2"
         units["length_unit"] = "au"
         units["force_unit"] = "mRy/au"
     elif interface_mode == "elk":
+        ElkToTHz = (
+            sqrt(physical_units.Hartree * physical_units.EV / physical_units.AMU)
+            / (physical_units.Bohr * 1e-10)
+            / (2 * pi)
+            / 1e12
+        )  # [THz] 154.10794
         units["factor"] = ElkToTHz
         units["nac_factor"] = 1.0
-        units["distance_to_A"] = Bohr
+        units["distance_to_A"] = physical_units.Bohr
         units["force_constants_unit"] = "hartree/au^2"
         units["length_unit"] = "au"
         units["force_unit"] = "hartree/au"
     elif interface_mode in ["siesta", "abacus"]:
+        SiestaToTHz = (
+            sqrt(physical_units.EV / (physical_units.AMU * physical_units.Bohr))
+            / physical_units.Angstrom
+            / (2 * pi)
+            / 1e12
+        )  # [THz] 21.49068
         units["factor"] = SiestaToTHz
-        units["nac_factor"] = Hartree / Bohr
-        units["distance_to_A"] = Bohr
+        units["nac_factor"] = physical_units.Hartree / physical_units.Bohr
+        units["distance_to_A"] = physical_units.Bohr
         units["force_constants_unit"] = "eV/angstrom.au"
         units["length_unit"] = "au"
         units["force_unit"] = "eV/angstrom"
     elif interface_mode == "cp2k":
+        CP2KToTHz = (  # CP2K uses a.u. for forces but Angstrom for distances
+            sqrt(
+                physical_units.Hartree
+                * physical_units.EV
+                / (physical_units.AMU * physical_units.Bohr)
+            )
+            / physical_units.Angstrom
+            / (2 * pi)
+            / 1e12
+        )
         units["factor"] = CP2KToTHz
-        units["nac_factor"] = None  # not implemented
+        units["nac_factor"] = physical_units.Bohr**2
         units["distance_to_A"] = 1.0
+        units["force_to_eVperA"] = physical_units.Hartree / physical_units.Bohr
         units["force_constants_unit"] = "hartree/angstrom.au"
         units["length_unit"] = "angstrom"
         units["force_unit"] = "hartree/au"
     elif interface_mode == "crystal":
+        CrystalToTHz = physical_units.DefaultToTHz
         units["factor"] = CrystalToTHz
-        units["nac_factor"] = Hartree * Bohr
+        units["nac_factor"] = physical_units.Hartree * physical_units.Bohr
         units["distance_to_A"] = 1.0
         units["force_constants_unit"] = "eV/angstrom^2"
         units["length_unit"] = "angstrom"
         units["force_unit"] = "eV/angstrom"
     elif interface_mode == "dftbp":
+        DftbpToTHz = (
+            sqrt(physical_units.Hartree * physical_units.EV / physical_units.AMU)
+            / (physical_units.Bohr * 1e-10)
+            / (2 * pi)
+            / 1e12
+        )  # [THz] 154.10794344
         units["factor"] = DftbpToTHz
-        units["nac_factor"] = Hartree * Bohr
-        units["distance_to_A"] = Bohr
+        units["nac_factor"] = physical_units.Hartree * physical_units.Bohr
+        units["distance_to_A"] = physical_units.Bohr
         units["force_constants_unit"] = "hartree/au^2"
         units["length_unit"] = "au"
         units["force_unit"] = "hartree/au"
     elif interface_mode == "turbomole":
+        TurbomoleToTHz = (  # Turbomole uses atomic units (Hartree/Bohr)
+            sqrt(physical_units.Hartree * physical_units.EV / physical_units.AMU)
+            / (physical_units.Bohr * 1e-10)
+            / (2 * pi)
+            / 1e12
+        )  # [THz] 154.10794
         units["factor"] = TurbomoleToTHz
         units["nac_factor"] = 1.0
-        units["distance_to_A"] = Bohr
-        units["force_to_eVperA"] = Hartree / Bohr
+        units["distance_to_A"] = physical_units.Bohr
+        units["force_to_eVperA"] = physical_units.Hartree / physical_units.Bohr
         units["force_constants_unit"] = "hartree/au^2"
         units["length_unit"] = "au"
         units["force_unit"] = "hartree/au"
     elif interface_mode == "castep":
+        CastepToTHz = physical_units.DefaultToTHz
         units["factor"] = CastepToTHz
-        units["nac_factor"] = Hartree * Bohr
+        units["nac_factor"] = physical_units.Hartree * physical_units.Bohr
         units["distance_to_A"] = 1.0
         units["force_constants_unit"] = "eV/angstrom^2"
         units["length_unit"] = "angstrom"
         units["force_unit"] = "eV/angstrom"
     elif interface_mode == "fleur":
+        FleurToTHz = (  # Fleur uses atomic units (Hartree/Bohr)
+            sqrt(physical_units.Hartree * physical_units.EV / physical_units.AMU)
+            / (physical_units.Bohr * 1e-10)
+            / (2 * pi)
+            / 1e12
+        )  # [THz] 154.10794
         units["factor"] = FleurToTHz
         units["nac_factor"] = 1.0
-        units["distance_to_A"] = Bohr
+        units["distance_to_A"] = physical_units.Bohr
         units["force_constants_unit"] = "hartree/au^2"
         units["length_unit"] = "au"
         units["force_unit"] = "hartree/au"
@@ -815,15 +877,16 @@ def get_calc_dataset_wien2k(
 def get_force_constant_conversion_factor(unit, interface_mode):
     """Return unit conversion factor of force constants."""
     _unit = unit.replace("Angstrom", "angstrom")  # for backward compatibility
-    interface_default_units = get_default_physical_units(interface_mode)
+    interface_default_units = get_calculator_physical_units(interface_mode)
     default_unit = interface_default_units["force_constants_unit"]
+    physical_units = get_physical_units()
     factor_to_eVperA2 = {
         "eV/angstrom^2": 1,
-        "eV/angstrom.au": 1 / Bohr,
-        "Ry/au^2": Rydberg / Bohr**2,
-        "mRy/au^2": Rydberg / Bohr**2 / 1000,
-        "hartree/au^2": Hartree / Bohr**2,
-        "hartree/angstrom.au": Hartree / Bohr,
+        "eV/angstrom.au": 1 / physical_units.Bohr,
+        "Ry/au^2": physical_units.Rydberg / physical_units.Bohr**2,
+        "mRy/au^2": physical_units.Rydberg / physical_units.Bohr**2 / 1000,
+        "hartree/au^2": physical_units.Hartree / physical_units.Bohr**2,
+        "hartree/angstrom.au": physical_units.Hartree / physical_units.Bohr,
     }
     if default_unit not in factor_to_eVperA2:
         msg = "Force constant conversion for %s unit is not implemented."