+++ phonopy/scripts/phonopy_gruneisen.py
@@ -39,8 +39,8 @@ import numpy as np
 import phonopy
 from phonopy import PhonopyGruneisen
 from phonopy.interface.calculator import (
+    get_calculator_physical_units,
     get_default_cell_filename,
-    get_default_physical_units,
     get_interface_mode,
 )
 
@@ -261,10 +261,10 @@ def run():
     # Phonopy gruneisen interface mode
     #
     interface_mode = get_interface_mode(vars(args))
-    physical_units = get_default_physical_units(interface_mode)
+    units = get_calculator_physical_units(interface_mode)
 
     if args.factor is not None:
-        physical_units["factor"] = args.factor
+        units["factor"] = args.factor
 
     if args.cell_filename:
         cell_filename = args.cell_filename
@@ -297,7 +297,7 @@ def run():
                 unitcell_filename=unitcell_filename,
                 born_filename=born_filename,
                 force_constants_filename=fc_filename,
-                factor=physical_units["factor"],
+                factor=units["factor"],
                 symprec=args.symprec,
             )
             phonons.append(phonon)
@@ -311,7 +311,7 @@ def run():
                 unitcell_filename=unitcell_filename,
                 born_filename=born_filename,
                 force_sets_filename=force_filename,
-                factor=physical_units["factor"],
+                factor=units["factor"],
                 symprec=args.symprec,
             )
             phonons.append(phonon)