+++ phonopy/spectrum/velocity.py
@@ -39,8 +39,8 @@ import sys
 import numpy as np
 
 from phonopy.harmonic.dynmat_to_fc import get_commensurate_points
+from phonopy.physical_units import get_physical_units
 from phonopy.structure.grid_points import get_qpoints
-from phonopy.units import AMU, kb_J
 
 
 class Velocity:
@@ -206,7 +206,11 @@ class AutoCorrelation:
         self._vv = vv
         if self._masses is not None and self._temperature is not None:
             for i, m in enumerate(self._masses):
-                self._vv[:, i] *= m * AMU / (kb_J * self._temperature)
+                self._vv[:, i] *= (
+                    m
+                    * get_physical_units().AMU
+                    / (get_physical_units().kb_J * self._temperature)
+                )
 
         self._n_elements = n_elem